project_bot/requirements.txt
aiogram
python-dotenv
motor
